[BEAM-3310] Extract CustomQueryResults to be used in all metrics sinks tests.
